function are now in their own class
